+++ src/node-saml/algorithms.ts
