+++ src/node-saml/saml-post-signing.ts
