+++ src/node-saml/utility.ts
